refactor(core): separate master password type and policy
