+++ packages/core/src/domain/master-password/index.ts
@@ -1,0 +1,1 @@
+export type { RawMasterPassword } from "./master-password.type";
