+++ packages/core/src/domain/master-password/master-password.type.ts
@@ -1,0 +1,3 @@
+import type { Brand } from "../common/brand-keys";
+
+export type RawMasterPassword = Brand<string, "RawMasterPassword">;
